+++ platform/ios/src/MGLMapView.h
@@ -226,6 +226,12 @@ IB_DESIGNABLE
 - (IBAction)reloadStyle:(id)sender;
 
 /**
+ A control indicating the scale of the map. The scale bar is positioned in the
+ upper-left corner. The scale bar is hidden by default.
+ */
+@property (nonatomic, readonly) UIView *scaleBar;
+
+/**
  A control indicating the map’s direction and allowing the user to manipulate
  the direction, positioned in the upper-right corner.
  */